requirements.txt
pyTelegramBotAPI
piecash
python-dotenv
mysql-connector-python
pyyaml
